refactor(examples): move training utilities out of models
