+++ docs/source/legacy_tutorials/cn/16_train_large_scale_snn.rst
@@ -403,91 +403,9 @@ ResNet-18的预训练模型参数。运行结果为：
 
 让网络自行复制，计算效率会稍微高一些，原因参见 :doc:`时间驱动：使用卷积SNN识别Fashion-MNIST <./4_conv_fashion_mnist>`。
 
-在ImageNet上训练
+在 ImageNet 上训练
 ---------------------------------------
-ImageNet [#ImageNet]_ 是计算机视觉常用的数据集，对于SNN而言颇具挑战性。惊蜇框架提供了一个训练ImageNet的代码样例，位于
-`spikingjelly.activation_based.model.train_imagenet <https://github.com/fangwei123456/spikingjelly/blob/master/spikingjelly/activation_based/model/train_imagenet.py>`_ 。
-该代码样例的实现参考了 `torchvision <https://github.com/pytorch/vision/blob/main/references/classification/train.py>`_ 。使用
-时只需要构建好网络、损失函数和正确率计算方式，就可以快速训练，下面是使用示例：
 
-.. code:: python
-
-    import torch
-    import torch.nn.functional as F
-    from spikingjelly.activation_based.model import train_imagenet, spiking_resnet, train_classify
-    from spikingjelly.activation_based import neuron, surrogate
-
-    def ce_loss(x_seq: torch.Tensor, label: torch.Tensor):
-        # x_seq.shape = [T, N, C]
-        return F.cross_entropy(input=x_seq.mean(0), target=label)
-
-    def cal_acc1_acc5(output, target):
-        return train_classify.default_cal_acc1_acc5(output.mean(0), target)
-
-
-    if __name__ == '__main__':
-        net = spiking_resnet.multi_step_spiking_resnet18(T=4, multi_step_neuron=neuron.MultiStepIFNode, surrogate_function=surrogate.ATan(), detach_reset=True, backend='cupy')
-        args = train_imagenet.parse_args()
-        train_imagenet.main(model=net, criterion=ce_loss, args=args, cal_acc1_acc5=cal_acc1_acc5)
-
-我们把这段代码保存为 `resnet18_imagenet.py`。查看运行参数：
-
-.. code:: shell
-
-    (pytorch-env) wfang@onebrain-dgx-a100-01:~/ssd/temp_dir$ python resnet18_imagenet.py -h
-
-                                [--step-gamma STEP_GAMMA] [--cosa-tmax COSA_TMAX] [--momentum M] [--wd W] [--output-dir OUTPUT_DIR] [--resume RESUME] [--start-epoch N] [--cache-dataset]
-                                [--sync-bn] [--amp] [--world-size WORLD_SIZE] [--dist-url DIST_URL] [--tb] [--T T] [--local_rank LOCAL_RANK]
-
-    PyTorch Classification Training
-
-    optional arguments:
-      -h, --help            show this help message and exit
-      --data-path DATA_PATH
-                            dataset
-      --device DEVICE       device
-      -b BATCH_SIZE, --batch-size BATCH_SIZE
-      --epochs N            number of total epochs to run
-      -j N, --workers N     number of data loading workers (default: 16)
-      --lr LR               initial learning rate
-      --opt OPT             optimizer (sgd or adam)
-      --lrs LRS             lr schedule (cosa(CosineAnnealingLR), step(StepLR)) or None
-      --step-size STEP_SIZE
-                            step_size for StepLR
-      --step-gamma STEP_GAMMA
-                            gamma for StepLR
-      --cosa-tmax COSA_TMAX
-                            T_max for CosineAnnealingLR. If none, it will be set to epochs
-      --momentum M          Momentum for SGD
-      --wd W, --weight-decay W
-                            weight decay (default: 0)
-      --output-dir OUTPUT_DIR
-                            path where to save
-      --resume RESUME       resume from checkpoint
-      --start-epoch N       start epoch
-      --cache-dataset       Cache the datasets for quicker initialization. It also serializes the transforms
-      --sync-bn             Use sync batch norm
-      --amp                 Use AMP training
-      --world-size WORLD_SIZE
-                            number of distributed processes
-      --dist-url DIST_URL   url used to set up distributed training
-      --tb                  Use TensorBoard to record logs
-      --T T                 simulation steps
-      --local_rank LOCAL_RANK
-
-
-在单卡上训练：
-
-.. code:: shell
-
-    python resnet18_imagenet.py --data-path /raid/wfang/datasets/ImageNet --lr 0.1 --opt sgd --lrs cosa --amp --tb --device cuda:7
-
-在多卡上训练：
-
-.. code:: shell
-
-    python -m torch.distributed.launch --nproc_per_node=8 resnet18_imagenet.py --data-path /raid/wfang/datasets/ImageNet --lr 0.1 --opt sgd --lrs cosa --amp --tb
+旧版训练 API 已移除。请参阅当前教程 :doc:`大规模深度SNN的训练 <../../tutorials/cn/train_large_scale_snn>`。
 
 .. [#ResNet] He, Kaiming, et al. "Deep residual learning for image recognition." Proceedings of the IEEE conference on computer vision and pattern recognition. 2016.
-
-.. [#ImageNet] Deng, Jia, et al. "Imagenet: A large-scale hierarchical image database." 2009 IEEE conference on computer vision and pattern recognition. IEEE, 2009.
