+++ docs/source/legacy_tutorials/en/15_recurrent_connection_and_stateful_synapse.rst
@@ -107,7 +107,6 @@ First, let us import packages:
     import torch.nn as nn
     import torch.nn.functional as F
     import torchvision.datasets
-    from spikingjelly.activation_based.model import train_classify
     from spikingjelly.activation_based import neuron, surrogate, layer
     from spikingjelly.activation_based.functional import seq_to_ann_forward
     from torchvision import transforms
