+++ spikingjelly/activation_based/examples/common/tv_ref_classify/utils.py
@@ -1,11 +1,8 @@
-import copy
 import datetime
-import errno
-import hashlib
 import os
 import statistics
 import time
-from collections import OrderedDict, defaultdict, deque
+from collections import defaultdict, deque
 from numbers import Number
 from typing import Optional, Sequence, Union
 
@@ -17,7 +14,6 @@ from spikingjelly.logger import logger
 __all__ = [
     "SmoothedValue",
     "MetricLogger",
-    "mkdir",
     "init_distributed_mode",
 ]
 
@@ -419,52 +415,8 @@ def accuracy(output, target, topk=(1,)):
         return res
 
 
-def mkdir(path):
-    r"""
-    **API Language** - :ref:`中文 <mkdir-cn>` | :ref:`English <mkdir-en>`
-
-    ----
-
-    .. _mkdir-cn:
-
-    * **中文**
-
-    创建目录。如果目录已存在则忽略。
-
-    :param path: 要创建的目录路径
-    :type path: str
-
-    ----
-
-    .. _mkdir-en:
-
-    * **English**
-
-    Create a directory. Ignore if it already exists.
-
-    :param path: Directory path to create
-    :type path: str
-    """
-
-    try:
-        os.makedirs(path)
-    except OSError as e:
-        if e.errno != errno.EEXIST:
-            raise
-
-
 def is_dist_avail_and_initialized():
-    if not dist.is_available():
-        return False
-    if not dist.is_initialized():
-        return False
-    return True
-
-
-def get_world_size():
-    if not is_dist_avail_and_initialized():
-        return 1
-    return dist.get_world_size()
+    return dist.is_available() and dist.is_initialized()
 
 
 def get_rank():
@@ -535,133 +487,6 @@ def init_distributed_mode(args):
     )
     if args.rank == 0:
         logger.info("Distributed training initialized: {}", args.dist_url)
-
-
-def average_checkpoints(inputs):
-    """Loads checkpoints from inputs and returns a model with averaged weights. Original implementation taken from:
-    https://github.com/pytorch/fairseq/blob/a48f235636557b8d3bc4922a6fa90f3a0fa57955/scripts/average_checkpoints.py#L16
-
-    Args:
-      inputs (List[str]): An iterable of string paths of checkpoints to load from.
-    Returns:
-      A dict of string keys mapping to various values. The 'model' key
-      from the returned dict should correspond to an OrderedDict mapping
-      string parameter names to torch Tensors.
-    """
-    params_dict = OrderedDict()
-    params_keys = None
-    new_state = None
-    num_models = len(inputs)
-    for fpath in inputs:
-        with open(fpath, "rb") as f:
-            state = torch.load(
-                f,
-                map_location=(
-                    lambda s, _: torch.serialization.default_restore_location(s, "cpu")
-                ),
-            )
-        # Copies over the settings from the first checkpoint
-        if new_state is None:
-            new_state = state
-        model_params = state["model"]
-        model_params_keys = list(model_params.keys())
-        if params_keys is None:
-            params_keys = model_params_keys
-        elif params_keys != model_params_keys:
-            raise KeyError(
-                f"For checkpoint {f}, expected list of params: {params_keys}, but found: {model_params_keys}"
-            )
-        for k in params_keys:
-            p = model_params[k]
-            if isinstance(p, torch.HalfTensor):
-                p = p.float()
-            if k not in params_dict:
-                params_dict[k] = p.clone()
-                # NOTE: clone() is needed in case of p is a shared parameter
-            else:
-                params_dict[k] += p
-    averaged_params = OrderedDict()
-    for k, v in params_dict.items():
-        averaged_params[k] = v
-        if averaged_params[k].is_floating_point():
-            averaged_params[k].div_(num_models)
-        else:
-            averaged_params[k] //= num_models
-    new_state["model"] = averaged_params
-    return new_state
-
-
-def store_model_weights(model, checkpoint_path, checkpoint_key="model", strict=True):
-    """
-    This method can be used to prepare weights files for new models. It receives as
-    input a model architecture and a checkpoint from the training script and produces
-    a file with the weights ready for release.
-
-    Examples:
-        from torchvision import models as M
-
-        # Classification
-        model = M.mobilenet_v3_large(pretrained=False)
-        logger.info("{}", store_model_weights(model, './class.pth'))
-
-        # Quantized Classification
-        model = M.quantization.mobilenet_v3_large(pretrained=False, quantize=False)
-        model.fuse_model(is_qat=True)
-        model.qconfig = torch.ao.quantization.get_default_qat_qconfig('qnnpack')
-        _ = torch.ao.quantization.prepare_qat(model, inplace=True)
-        logger.info("{}", store_model_weights(model, './qat.pth'))
-
-        # Object Detection
-        model = M.detection.fasterrcnn_mobilenet_v3_large_fpn(pretrained=False, pretrained_backbone=False)
-        logger.info("{}", store_model_weights(model, './obj.pth'))
-
-        # Segmentation
-        model = M.segmentation.deeplabv3_mobilenet_v3_large(pretrained=False, pretrained_backbone=False, aux_loss=True)
-        logger.info("{}", store_model_weights(model, './segm.pth', strict=False))
-
-    Args:
-        model (pytorch.nn.Module): The model on which the weights will be loaded for validation purposes.
-        checkpoint_path (str): The path of the checkpoint we will load.
-        checkpoint_key (str, optional): The key of the checkpoint where the model weights are stored.
-            Default: "model".
-        strict (bool): whether to strictly enforce that the keys
-            in :attr:`state_dict` match the keys returned by this module's
-            :meth:`~torch.nn.Module.state_dict` function. Default: ``True``
-
-    Returns:
-        output_path (str): The location where the weights are saved.
-    """
-    # Store the new model next to the checkpoint_path
-    checkpoint_path = os.path.abspath(checkpoint_path)
-    output_dir = os.path.dirname(checkpoint_path)
-
-    # Deep copy to avoid side-effects on the model object.
-    model = copy.deepcopy(model)
-    checkpoint = torch.load(checkpoint_path, map_location="cpu")
-
-    # Load the weights to the model to validate that everything works
-    # and remove unnecessary weights (such as auxiliaries, etc)
-    if checkpoint_key == "model_ema":
-        del checkpoint[checkpoint_key]["n_averaged"]
-        torch.nn.modules.utils.consume_prefix_in_state_dict_if_present(
-            checkpoint[checkpoint_key], "module."
-        )
-    model.load_state_dict(checkpoint[checkpoint_key], strict=strict)
-
-    tmp_path = os.path.join(output_dir, str(model.__hash__()))
-    torch.save(model.state_dict(), tmp_path)
-
-    sha256_hash = hashlib.sha256()
-    with open(tmp_path, "rb") as f:
-        # Read and update hash string value in blocks of 4K
-        for byte_block in iter(lambda: f.read(4096), b""):
-            sha256_hash.update(byte_block)
-        hh = sha256_hash.hexdigest()
-
-    output_path = os.path.join(output_dir, "weights-" + str(hh[:8]) + ".pth")
-    os.replace(tmp_path, output_path)
-
-    return output_path
 
 
 def reduce_across_processes(
